// speedyGo-ui/src/app/services/models/station.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Station {
  capacity?: number;
  createdAt?: string;
  id?: number;
  is_active?: boolean;
  latitude?: number;
  location?: string;
  longitude?: number;
  name?: string;
  updatedAt?: string;
  available_vehicles?:number ;

}